+++ rpc-client/src/mock_sender.rs
@@ -397,6 +397,7 @@ impl RpcSender for MockSender {
                 rpc: Some(SocketAddr::from(([10, 239, 6, 48], 8899))),
                 pubsub: Some(SocketAddr::from(([10, 239, 6, 48], 8900))),
                 version: Some("1.0.0 c375ce1f".to_string()),
+                client_id: Some(0),
                 feature_set: None,
                 shred_version: None,
             }])?,